requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.101.1
fastapi-pagination==0.12.8
h11==0.14.0
idna==3.4
iso8601==1.1.0
passlib==1.7.4
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT==2.8.0
PyPika==0.48.9
pypika-tortoise==0.1.6
python-multipart==0.0.6
pytz==2023.3
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.7.1
uvicorn==0.23.2
websockets==11.0.3
